requirements.txt
django
pymysql
pyjade
django-tinymce
markdown
django-filter
django-ordered-model
django-suit
pillow
djangorestframework
djangorestframework-csv
python-dateutil
requests
django-password-reset
easy-thumbnails
django-object-actions
drf-extensions